+++ builder/kubernetes/builder_test.go
@@ -1,4 +1,4 @@
-package builder_k8s
+package kubernetes
 
 import (
 	"context"
