+++ builder/kubernetes/testdata/Dockerfile
