+++ builder/kubernetes/testdata/greetings.sw.json
